Move test tc to seperate file.
